pyproject.toml
[tool.poetry]
name = "pws"
version = "0.1.0"
description = "PWS AutoTest"
authors = ["Oleksandr Voldieiev"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10.2"
allure-pytest = "^2.13.2"
ipdb = "^0.13.9"
pytest = "^7.3.2"
ipdbplugin = "^1.5.0"
PyYAML = "^6.0"
pytest-lazy-fixture = "^0.6.3"
pytest-playwright = "^0.4.2"
jinja2 = "^3.1.2"
mypy = "^1.0.0"
dataclasses-json = "^0.5.7"
vulture = "^2.7"
pytest-rerunfailures = "^11.1.2"
python-dotenv = "^1.0.0"
vobject = "^0.9.6.1"
pillow = "^10.0.0"
pytest-test-groups = "^1.0.3"
pytest-timeout = "^2.1.0"
pytest-asyncio = "^0.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[pytest]
addopts = "--capture=no"
testpaths = "tests"
